--- a/lychee-bin/src/client.rs
+++ b/lychee-bin/src/client.rs
@@ -35,12 +35,21 @@ pub(crate) fn create(cfg: &Config, cookie_jar: Option<&Arc<CookieStoreMutex>>) -
     let rate_limit_config =
         RateLimitConfig::from_options(cfg.default_host_concurrency, cfg.default_request_interval);
     let cache_max_age = if cfg.cache { 3600 } else { 0 }; // 1 hour if caching enabled, disabled otherwise
-    let host_pool = HostPool::new(
-        rate_limit_config,
-        cfg.hosts.clone(),
-        cfg.max_concurrency,
-        cache_max_age,
-    );
+    let host_pool = match cookie_jar {
+        Some(cookie_jar) => HostPool::with_cookie_jar(
+            rate_limit_config,
+            cfg.hosts.clone(),
+            cfg.max_concurrency,
+            cache_max_age,
+            cookie_jar.clone(),
+        ),
+        None => HostPool::new(
+            rate_limit_config,
+            cfg.hosts.clone(),
+            cfg.max_concurrency,
+            cache_max_age,
+        ),
+    };
 
     ClientBuilder::builder()
         .remaps(remaps)