--- a/lychee-lib/src/extract/html/html5gum.rs
+++ b/lychee-lib/src/extract/html/html5gum.rs
@@ -249,7 +249,6 @@ impl LinkExtractor {
         if let Some(name) = self.current_attributes.get("name") {
             self.fragments.insert(name.to_string());
         }
-
         self.current_attributes.clear();
     }
 }