@mre 2025-09-07
Looks like crates.io has made some changes. I'll rerun it tomorrow to see if the issue remains.