@mre 2025-09-07
## 🐛 Bug Fix: File URL Handling

**Fixed**: File URLs () were incorrectly being processed through the rate limiting system, causing debug errors like:


**Solution**: File URLs are now properly excluded from rate limiting tracking since they're local filesystem operations, not network requests.

## 📖 Configuration Documentation

### Per-Host Configuration via Config File

The per-host concurrency and other settings can be customized in your  config file:

```toml
# Global defaults
default_host_concurrency = 10
default_request_interval = "100ms"

# Per-host overrides
[hosts."github.com"]
concurrency = 5          # Lower concurrency for GitHub API
request_interval = "200ms"  # More conservative timing

[hosts."api.example.com"]
concurrency = 1          # Single request at a time
request_interval = "1s"     # Very conservative for sensitive APIs
```

### CLI Options

```bash
# Set global defaults
lychee --default-host-concurrency 5 --default-request-interval 200ms

# Show per-host statistics
lychee --host-stats docs/

# Combined usage
lychee --default-host-concurrency 5 --host-stats --format json docs/
```

This provides flexible control over rate limiting behavior both globally and per specific hosts.