@mre 2025-09-24
Just so that I don't forget, I will need to fix the behavior when setting the request interval to 0. I'm getting some parsing errors.