@thomas-zahner 2025-10-03
@mre I seem to have a considerable performance regression with this PR. This might be related to the concurrency changes and maybe also with my comment [here](https://github.com/lycheeverse/lychee/pull/1844#discussion_r2360356264)?

I haven't done any through analysis or anything but I noticed it just by running lychee on the README and wanted to let you know. I switched multiple times between master and this PR and the performance degradation is very noticeable.

## master branch

```
➜  lychee git:(master) time ./target/debug/lychee README.md
   [WARN ] Error creating request: InvalidPathToUri("/CONTRIBUTING.md")
  111/111 ━━━━━━━━━━━━━━━━━━━━ Finished extracting links                                         🔍 111 Total (in 1s) ✅ 104 OK 🚫 0 Errors 🔀 7 Redirects
./target/debug/lychee README.md  0.36s user 0.06s system 22% cpu 1.836 total
➜  lychee git:(master) time ./target/debug/lychee README.md
   [WARN ] Error creating request: InvalidPathToUri("/CONTRIBUTING.md")
  111/111 ━━━━━━━━━━━━━━━━━━━━ Finished extracting links                                         🔍 111 Total (in 0s) ✅ 104 OK 🚫 0 Errors 🔀 7 Redirects
./target/debug/lychee README.md  0.35s user 0.05s system 58% cpu 0.671 total
➜  lychee git:(master) time ./target/debug/lychee README.md
   [WARN ] Error creating request: InvalidPathToUri("/CONTRIBUTING.md")
  111/111 ━━━━━━━━━━━━━━━━━━━━ Finished extracting links                                         🔍 111 Total (in 1s) ✅ 104 OK 🚫 0 Errors 🔀 7 Redirects
./target/debug/lychee README.md  0.39s user 0.06s system 28% cpu 1.594 total
```

## this PR

Sometimes I even felt like the link check process was stuck in the end.

```
➜  lychee git:(feat/per-host-rate-limiting) time ./target/debug/lychee README.md
   [WARN ] Error creating request: InvalidPathToUri("/CONTRIBUTING.md")
  111/111 ━━━━━━━━━━━━━━━━━━━━ Finished extracting links                                                                                                                                                                                🔍 111 Total (in 7s) ✅ 111 OK 🚫 0 Errors
./target/debug/lychee README.md  1.72s user 0.12s system 23% cpu 7.689 total
➜  lychee git:(feat/per-host-rate-limiting) time ./target/debug/lychee README.md
   [WARN ] Error creating request: InvalidPathToUri("/CONTRIBUTING.md")
  111/111 ━━━━━━━━━━━━━━━━━━━━ Finished extracting links                                                                                                                                                                                🔍 111 Total (in 6s) ✅ 111 OK 🚫 0 Errors
./target/debug/lychee README.md  1.72s user 0.10s system 26% cpu 6.825 total
➜  lychee git:(feat/per-host-rate-limiting) time ./target/debug/lychee README.md
   [WARN ] Error creating request: InvalidPathToUri("/CONTRIBUTING.md")
  111/111 ━━━━━━━━━━━━━━━━━━━━ Finished extracting links                                                                                                                                                                                🔍 111 Total (in 7s) ✅ 111 OK 🚫 0 Errors
./target/debug/lychee README.md  1.73s user 0.12s system 25% cpu 7.358 total
```